+++ android/src/main/java/com/magicleap/magicscript/scene/nodes/audio/AudioNode.kt
@@ -22,7 +22,6 @@ import com.facebook.react.bridge.ReadableMap
 import com.magicleap.magicscript.scene.nodes.audio.model.SpatialSoundDistance
 import com.magicleap.magicscript.scene.nodes.audio.model.SpatialSoundPosition
 import com.magicleap.magicscript.scene.nodes.base.TransformNode
-import com.magicleap.magicscript.utils.AudioFileProvider
 import com.magicleap.magicscript.utils.PropertiesReader.Companion.readFilePath
 import com.magicleap.magicscript.utils.ifContains
 import com.magicleap.magicscript.utils.putDefault
