+++ android/src/main/java/com/magicleap/magicscript/scene/nodes/audio/UriAudioProvider.kt
@@ -14,17 +14,19 @@
  *   limitations under the License.
  */
 
-package com.magicleap.magicscript.utils
+package com.magicleap.magicscript.scene.nodes.audio
 
 import android.content.Context
 import android.net.Uri
 import android.webkit.URLUtil
+import com.magicleap.magicscript.utils.logMessage
 import java.io.BufferedInputStream
 import java.io.File
 import java.io.InputStream
 import java.net.URL
 
-class UriAudioProvider(private val context: Context) : AudioFileProvider {
+class UriAudioProvider(private val context: Context) :
+    AudioFileProvider {
 
     companion object {
         private const val READ_TIMEOUT = 5000
@@ -33,61 +35,49 @@ class UriAudioProvider(private val context: Context) : AudioFileProvider {
 
     private val threads = mutableMapOf<String, Thread>()
 
-    override fun provideFile(uri: Uri?, result: (File) -> Unit): Boolean {
-        try {
-            if (uri == null) {
-                return false
-            }
-
-            val thread = Thread(
-                Runnable {
-                    try {
-                        var inputStream: InputStream =
-                            if (uri.toString().startsWith("android.resource")) {
-                                getLocalInputStream(uri)
-                            } else {
-                                getRemoteInputStream(uri)
-                            }
-
-                        val bufferedInputStream = BufferedInputStream(inputStream, 1024 * 5)
-
-                        val internalStorage = context.getDir("filesdir", Context.MODE_PRIVATE)
-                        val guessFileName = URLUtil.guessFileName(uri.toString(), null, null)
-                        val file = File("$internalStorage/$guessFileName")
-
-                        if (file.exists()) {
-                            file.delete()
+    override fun provideFile(uri: Uri, result: (File) -> Unit) {
+        val thread = Thread(
+            Runnable {
+                try {
+                    val inputStream: InputStream =
+                        if (uri.toString().startsWith("android.resource")) {
+                            getLocalInputStream(uri)
+                        } else {
+                            getRemoteInputStream(uri)
                         }
-                        file.createNewFile()
 
-                        file.copyInputStreamToFile(bufferedInputStream)
+                    val bufferedInputStream = BufferedInputStream(inputStream, 1024 * 5)
 
-                        bufferedInputStream.close()
-                        fileDownloaded(uri.path, result, file)
-                    } catch (e: Exception) {
-                        logMessage("Error during reading Audio File $e", true)
+                    val internalStorage = context.filesDir
+                    val guessFileName = URLUtil.guessFileName(uri.toString(), null, null)
+                    val file = File("$internalStorage/$guessFileName")
+
+                    if (file.exists()) {
+                        file.delete()
                     }
+                    file.createNewFile()
+
+                    file.copyInputStreamToFile(bufferedInputStream)
+
+                    bufferedInputStream.close()
+                    fileDownloaded(uri.path, result, file)
+                } catch (e: Exception) {
+                    logMessage("Error during reading Audio File $e", true)
                 }
-            )
-            threads[uri.path] = thread
-            thread.start()
-
-        } catch (e: Exception) {
-            logMessage("Error during reading Audio File $e", true)
-            return false
-        }
-
-        return true
+            }
+        )
+        threads[uri.path] = thread
+        thread.start()
     }
 
     private fun getRemoteInputStream(uri: Uri): InputStream {
         val url = URL(uri.toString())
-        val ucon = url.openConnection().apply {
+        val connection = url.openConnection().apply {
             readTimeout = READ_TIMEOUT
             connectTimeout = CONNECT_TIMEOUT
         }
 
-        return ucon.getInputStream()
+        return connection.getInputStream()
     }
 
     private fun getLocalInputStream(uri: Uri): InputStream {
@@ -113,11 +103,7 @@ class UriAudioProvider(private val context: Context) : AudioFileProvider {
         }
     }
 
-    private fun fileDownloaded(
-        path: String,
-        result: (File) -> Unit,
-        file: File
-    ) {
+    private fun fileDownloaded(path: String, result: (File) -> Unit, file: File) {
         result(file)
         threads[path]?.interrupt()
         threads.remove(path)
