+++ android/src/test/java/com/magicleap/magicscript/scene/nodes/audio/AudioNodeTest.kt
@@ -20,7 +20,6 @@ import androidx.test.core.app.ApplicationProvider
 import com.google.ar.sceneform.math.Vector3
 import com.magicleap.magicscript.*
 import com.magicleap.magicscript.scene.nodes.audio.model.SpatialSoundDistance
-import com.magicleap.magicscript.utils.UriAudioProvider
 import com.nhaarman.mockitokotlin2.*
 import org.amshove.kluent.shouldBe
 import org.amshove.kluent.shouldEqual
@@ -54,7 +53,6 @@ class AudioNodeTest {
                 val argument = it.arguments[1]
                 val result = argument as ((File) -> Unit)
                 result.invoke(downloadedFile)
-                true
             }
         }
